requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
sqlalchemy-utils
pgvector
pydantic
PyPDF2
pytesseract
Pillow
PyMuPDF
openai
pikepdf
